#71 💷 refactor: 컴포넌트 폴더 내에서 페이지 별로 폴더 생성 및 파일 위치 변경
