+++ src/components/exchange/Trading.tsx
